Rename exceptions file to errors.rb
