+++ lib/honey_format.rb
@@ -1,5 +1,5 @@
 require 'honey_format/version'
-require 'honey_format/exceptions'
+require 'honey_format/errors'
 require 'honey_format/csv'
 
 # Main module for HoneyFormat
