+++ lib/honey_format/errors.rb
